Fix: fallback exception handling Fix: fallback exception handling for molgenis-fair
